+++ src/OpenTelemetry.Resources.Host/CHANGELOG.md
@@ -13,6 +13,10 @@
 * Updated OpenTelemetry core component version(s) to `1.19.0`.
   ([#5240](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5240))
 
+* Added the `host.cpu.*` resource attributes, emitted when
+  `OTEL_DOTNET_EXPERIMENTAL_HOST_RESOURCE_ENABLE_CPU_INFO` is `true`.
+  ([#5173](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5173))
+
 ## 1.18.0-beta.1
 
 Released 2026-Aug-21